renamed module service to update_service_module
